wrp: first package persistence
This package offers a Context persistence object which basically deals with
adapters objects.

As of now we supply interaction with mongo nosql database through an adapter.
It is also included a mock pseudo database adapter.

Which adapter is loaded shall be predefined by the connection uri passed to
the supplier object.

Signed-off-by: Bob Ross <pianodaemon@gmail.com>